+++ src/app/(LowerPages)/career-path/[page]/CareerPath.module.css
